+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.17] - 2026-02-09
+
 ### Fixed
 
 - **Header Color Palette UX**: Improved header rename editor palette interactions and visuals